+++ src/main/java/de/samply/file/csv/CsvReader.java
@@ -1,0 +1,15 @@
+package de.samply.file.csv;
+
+import java.util.Iterator;
+import java.util.function.Consumer;
+import org.apache.commons.csv.CSVRecord;
+
+public class CsvReader {
+
+  private void doSomething(){
+
+  }
+
+
+
+}